src/app/home/home.component.ts
import { Component, OnInit } from '@angular/core';
import { FormBuilder, FormGroup } from '@angular/forms'

@Component({
  selector: 'app-home',
  templateUrl: './home.component.html',
  styleUrls: ['./home.component.sass']
})
export class HomeComponent implements OnInit {

  public form: FormGroup
  public name: string = 'World'

  constructor (
    private formBuilder: FormBuilder,
  ) {}

  ngOnInit() {
    this.form = this.formBuilder.group({
      name: []
    })
  }

  displayName() {
    console.log(this.form.value.name);
    this.name = this.form.value.name
  }

}
